[test] Fix windows empty path exists error
Windows fails with a TypeError when handing an empty
variable to os.path.exists().

+++ test/test_dtypes_integration.py
@@ -29,7 +29,7 @@ class TestTypesIntegration(unittest.TestCase):
         self.doc = doc
 
     def tearDown(self):
-        if os.path.exists(self.tmp_dir):
+        if self.tmp_dir and os.path.exists(self.tmp_dir):
             shutil.rmtree(self.tmp_dir)
 
     def test_time(self):
